+++ packages/commercetools/api-client/src/api/updateMyShoppingList/defaultMutation.ts
@@ -4,14 +4,46 @@ export default gql`
 
   mutation updateShoppingList($id: String!, $version: Long!, $actions: [MyShoppingListUpdateAction!]!, $locale: Locale!, $acceptLanguage: [Locale!]) {
     shoppingList: updateMyShoppingList(id: $id, version: $version, actions: $actions) {
-      id
-      version
+      id  
+      name(acceptLanguage: $acceptLanguage) 
+      description
+      slug
       lineItems {
-        id
+        id 
         productId
         variantId
         quantity
-        variant
+        variant {
+          id
+          sku
+          price (currency: "USD") {
+            tiers {
+              value {
+                centAmount
+              }
+            }
+            value {
+              centAmount
+            }
+            discounted {
+              value {
+                centAmount
+              }
+              discount {
+                isActive
+                name(acceptLanguage: $acceptLanguage)
+              }
+            }
+          }
+          images {
+            url
+            label
+          }
+        }
+      }
+      textLineItems {
+        name
+        description       
       }
     }
   }