+++ packages/commercetools/composables/src/useWishlist/currentWishlist.ts
@@ -1,8 +1,8 @@
 import { createMyShoppingList, getMyShoppingList } from '@vue-storefront/commercetools-api';
 
-const loadCurrentShoppingList = async (customQueryFn = (user = null) => ({user })) => {
-  const { user } = customQueryFn();
-  const { data: profileData } = await getMyShoppingList({ customer: false }, user);
+const loadCurrentShoppingList = async (customQueryFn = (wishlist = null) => ({ wishlist })) => {
+  const { wishlist } = customQueryFn();
+  const { data: profileData } = await getMyShoppingList({ customer: false });
 
   if (profileData.me.activeCart) {
     return profileData.me.activeCart;